feat(integrations): extract Resend into @nexaas/email-provider-resend (#88 Phase 2)
First reference integration to land under the new layout. Validates the
pattern; Postmark and SendGrid follow mechanically in subsequent PRs.

What moves:

- mcp/servers/email-outbound/src/providers/resend.ts ->
  integrations/email-provider-resend/src/index.ts
  (Git tracks as a rename; behavior unchanged.)

- ResendProvider class kept; new createResendProvider() factory exported
  alongside it as the stable surface for the Phase 3 manifest-driven
  loader to consume via dynamic import.

What gets pulled into the SDK:

- @nexaas/integration-sdk now exports the capability-specific shapes
  (EmailProvider, SendInput, SendOutput, TrackOutput) at
  src/capabilities/email-outbound.ts. Re-exported flat from index for
  now; can be re-namespaced to subpath exports later if more
  capabilities make the namespace crowded.

Shell wiring (mcp/servers/email-outbound):

- Now a workspace package (added to root workspaces array as a specific
  path so palace/memory MCP servers stay untouched).
- Depends on @nexaas/integration-sdk and @nexaas/email-provider-resend.
- src/types.ts becomes a thin re-export from @nexaas/integration-sdk so
  Postmark and SendGrid keep compiling without changes; they switch to
  importing directly from the SDK when each gets extracted.
- provider-select.ts swaps `new ResendProvider(key)` for
  `createResendProvider(key)` from the new package — one line per
  call site.

Changeset adds @nexaas/email-provider-resend at 0.1.0 (initial release).
The fixed group (framework packages) is unaffected; this is the first
independently-versioned integration.

No behavioral change to any running shell or skill: same provider class,
same selection logic, same env-var probe order. Only the package
boundary moved.

Stacks on #89 (feat/integrations-foundation-88).
