+++ .changeset/initial-resend-extraction.md
@@ -1,0 +1,5 @@
+---
+"@nexaas/email-provider-resend": minor
+---
+
+Initial release. Resend implementation of the `email-outbound` capability, migrated from `mcp/servers/email-outbound/src/providers/resend.ts`. Capability compat: `>=0.2 <1`.
