+++ mcp/servers/email-outbound/package.json
@@ -6,6 +6,8 @@
   "main": "src/index.ts",
   "dependencies": {
     "@modelcontextprotocol/sdk": "^1.0.0",
+    "@nexaas/integration-sdk": "0.1.0",
+    "@nexaas/email-provider-resend": "0.1.0",
     "zod": "^3.23.0"
   },
   "devDependencies": {
